Relax versions of hex.pm dependencies
Right now ltest does not work on any project trying to use lfe 2.1.3.
Since hex.pm requires semver, and rebar3 supports this, we can relax the
required versions listed in the rebar.config. This allows for this
library to work with any minor version without requiring a version bump
for ltest.